stock_anamoly_detection/validator/requirements.txt
pydantic
quixstreams
tqdm
python-dotenv
pandas
zstandard
